// ResuMe/ResuMe.WebApp/src/app/models/user.ts
export class User{
    email: string;
    userName: string;
    fullName: string;

    curriculum: any;

    profileImage: string;

    private: boolean;
    currentUser :boolean;

    linkedin: string;
    currentUserName: string;

    yearsOfExperience: number;
    role: string;
    birthdate: string;
    years: number;
}